Add files via upload
This work is about BitLocker Drive Encryption. We are going to install the feature using Power Shell, and create a GPO to deploy the feature to the other computers in the domain. In addition to this, we explain how to activate it on the client, and how to create a GPO to store the recovery key in Active Directory. Enjoy it!